bot/requirements.txt
aiogram==3.4.1
geopy==2.3.0
python-dotenv==1.0.1
python_dateutil==2.8.2
pytz==2024.1
timezonefinder==6.5.0
